create a repository
Black-box™ is a game of "hide and seek", which simulates shooting rays
into a black box to deduce the locations of mirrors hidden inside. It
was created by Eric Solomon in 1970s. Currently it has numerous computer
implementations for many different platforms, perhaps this one is the
first on Windows Phone platform,  hope you like it.